+++ tanaguru-engine/tanaguru-api/src/main/java/com/tanaguru/domain/dto/ProjectDTO.java
@@ -1,12 +1,15 @@
 package com.tanaguru.domain.dto;
 
+import javax.validation.constraints.Max;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Size;
 
 public class ProjectDTO {
 
     @NotNull
     @NotEmpty
+    @Size(max = 30, min = 3)
     private String name;
 
     private long contractId;