+++ tanaguru-engine/tanaguru-api/src/main/java/com/tanaguru/repository/TestHierarchyResultRepository.java
@@ -18,5 +18,6 @@ public interface TestHierarchyResultRepository extends JpaRepository<TestHierarc
     Optional<TestHierarchyResult> findByTestHierarchyAndPage(TestHierarchy testHierarchy, Page page);
     Collection<TestHierarchyResult> findAllByPage(Page page);
     Collection<TestHierarchyResult> findAllByPage_InAndTestHierarchy(Collection<Page> pages, TestHierarchy testHierarchy);
+    Collection<TestHierarchyResult> findAllByPageAndTestHierarchy(Page page, TestHierarchy testHierarchy);
     Collection<TestHierarchyResult> findAllByPage_AuditAndTestHierarchy(Audit audit, TestHierarchy testHierarchy);
 }