+++ CHANGELOG.md
@@ -5,5 +5,8 @@
 - User, contract and project management
 
 ## 1.0.2
+- Endpoint healthcheck , new key in common properties : management.endpoint.health.show-details, in order to show details or not
 - Rest : New key "admin.mail.whenblocked" (Allow send an email to admin when user is blocked)
+- JSON export of audits and pages
+- Send email at the end of the audit (site, scenario, group of pages)
 - Possibility to check the loading of a page with the search of a css selector